requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aiohttp=3.8.1=pypi_0
aiosignal=1.3.1=pypi_0
async-timeout=4.0.2=pypi_0
attrs=23.1.0=pypi_0
blas=1.0=mkl
brotlipy=0.7.0=py310h2bbff1b_1002
bzip2=1.0.8=he774522_0
ca-certificates=2023.01.10=haa95532_0
certifi=2023.5.7=py310haa95532_0
cffi=1.15.1=py310h2bbff1b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
colorama=0.4.6=pypi_0
contourpy=1.0.7=pypi_0
cryptography=39.0.1=py310h21b164f_0
cycler=0.11.0=pypi_0
einops=0.6.0=pypi_0
filelock=3.12.0=pypi_0
fonttools=4.39.4=pypi_0
freetype=2.12.1=ha860e81_0
frozenlist=1.3.3=pypi_0
fsspec=2023.5.0=pypi_0
giflib=5.2.1=h8cc25b3_3
huggingface-hub=0.14.1=pypi_0
idna=3.4=py310haa95532_0
imutils=0.5.4=pypi_0
intel-openmp=2021.4.0=haa95532_3556
jpeg=9e=h2bbff1b_1
kiwisolver=1.4.4=pypi_0
lerc=3.0=hd77b12b_0
libdeflate=1.17=h2bbff1b_0
libffi=3.4.4=hd77b12b_0
libpng=1.6.39=h8cc25b3_0
libtiff=4.5.0=h6c2663c_2
libuv=1.44.2=h2bbff1b_0
libwebp=1.2.4=hbc33d0d_1
libwebp-base=1.2.4=h2bbff1b_1
lz4-c=1.9.4=h2bbff1b_0
matplotlib=3.7.1=pypi_0
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py310h2bbff1b_0
mkl_fft=1.3.1=py310ha0764ea_0
mkl_random=1.2.2=py310h4ed8f06_0
multidict=6.0.4=pypi_0
numpy=1.23.4=py310h60c9a35_0
numpy-base=1.23.4=py310h04254f7_0
opencv-python=4.6.0.66=pypi_0
openssl=1.1.1t=h2bbff1b_0
packaging=23.1=pypi_0
pandas=2.0.1=pypi_0
pillow=9.4.0=py310hd77b12b_0
pip=22.3.1=py310haa95532_0
pycparser=2.21=pyhd3eb1b0_0
pymodbus=3.2.2=pypi_0
pyopenssl=23.0.0=py310haa95532_0
pyparsing=3.0.9=pypi_0
pysocks=1.7.1=py310haa95532_0
python=3.10.8=h966fe2a_1
python-dateutil=2.8.2=pypi_0
pytorch=1.13.0=py3.10_cpu_0
pytorch-mutex=1.0=cpu
pytz=2023.3=pypi_0
pyyaml=6.0=pypi_0
requests=2.29.0=py310haa95532_0
scipy=1.10.1=pypi_0
seaborn=0.12.2=pypi_0
setuptools=65.7.0=pypi_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h2bbff1b_0
timm=0.6.12=pypi_0
tk=8.6.12=h2bbff1b_0
torchvision=0.14.0=py310_cpu
tqdm=4.65.0=pypi_0
typing_extensions=4.5.0=py310haa95532_0
tzdata=2023.3=pypi_0
urllib3=1.26.15=py310haa95532_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.38.4=py310haa95532_0
win_inet_pton=1.1.0=py310haa95532_0
xz=5.4.2=h8cc25b3_0
yarl=1.9.2=pypi_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
